+++ src/livestreamer_cli/argparser.py
@@ -260,6 +260,13 @@ general.add_argument(
     """
 )
 general.add_argument(
+    "--version-check",
+    action="store_true",
+    help="""
+    Runs a version check and exits.
+    """
+)
+general.add_argument(
     "--yes-run-as-root",
     action="store_true",
     help=argparse.SUPPRESS